Removed size_t from libBitIO in favor of uint64_t

+++ libBitIO/include/BWSTransform.h
@@ -11,7 +11,7 @@ extern "C" {
     
     void EncodeBWSTransform(uint16_t ****Array, size_t ArraySize);
     
-    void RotateArray(const size_t DataSize, int64_t *Data, const uint64_t NumRotations, const bool RotateRight);
+    void RotateArray(const uint64_t DataSize, int64_t *Data, const uint64_t NumRotations, const bool RotateRight);
     
 #ifdef __cplusplus
 }
